Create 闭哈希表模板
Description
You need to insert integer keys into a closed hash table with n slots. The hash function used here is H(k) = k mod n, and the linear probing function is used for collision resolution.

Input
The first line, n,  is the length of the hash table, and the second line is the sequence of integer keys for insertion in the hash table.

 

Output
Output the keys stored in the hash table in order.  Each line contains two inegers, the first one of which is the slot number and the second one is the  key stored in the slot.  
 

Sample Input
13
12 23 45 57 20 3 78 31 15 36

Sample Output
0 78
2 15
3 3
5 57
6 45
7 20
8 31
10 23
11 36
12 12