Create 判断是否为平衡树
题目为：
Description
A 2-tree is a tree in which each node either has exactly 2 non-empty children (its left and right child), or is a leaf (has no children). Each leaf is named with an uppercase letter, and each other node is named with a lowercase letter. We want to know whether a given 2-tree is a balance tree. If a 2-tree satisfies the following conditions, we say it is a balance tree:（1）If the height of the left sub-tree of the root is defined as h1 and the height of the right sub-tree of the root is defined as h2, |h1 – h2| <= 1.（2）The left sub-tree and right sub-tree of the root are both balance trees. 

Input
The input consists of T test cases. The number of test cases (T) is given in the first line of the input. Each one of the following T lines contains a string representing the preorder traversal sequence of a 2-tree, each character in which is a letter ('A'-'Z','a'-'z').

Output
 The output consists of T lines. If the No. i tree in the input is a balance tree, ‘B’ is printed in the NO.i line of the output. If it is not a balance tree, ‘N’ is printed. 

Sample Input
4

qxABzCgRQ

rAB

bCdEfGH 

abmqPDWSE

Sample Output
B

B

N

N

HINT
Tips：You can write a recursive function to construct the tree from a string such as ‘qxABzCgRQ‘. The declaration of the function can be as follows:

//str – the string containing the characters of nodes.

//nlen – the length of the string str

//root – the returned parameter to record the pointer pointing to the root of the constructed tree

//nsize – the returned parameter to record the node count of the constructed tree



void LoadTree (char* str,  int nlen,  Node*& root,  int &nsize);

{

… …

//construct the root.

… …

if(the root is an internal node)

{

   … …

//construct the left sub-tree

LoadTree (str + 1, nlen – 1, root->left, size);

nsize += size;



//construct the right sub-tree

LoadTree(str+1+size, nlen – 1 – size, root->right, size)

nsize += size;
}

       else //the root is a leaf

{

.. …
}

}